BUG: ICU4N.Globalization.UChar::IsWhiteSpace(char) - method was calling itself, added a cast to int

+++ src/ICU4N/Globalization/UCharacter.cs
@@ -879,7 +879,7 @@ namespace ICU4N
         // ICU4N TODO: Tests
         public static bool IsWhiteSpace(char c) // ICU4N specific overload to cover System.Char
         {
-            return IsWhiteSpace(c);
+            return IsWhiteSpace((int)c);
         }
 
         /// <summary>
